sendNotif in lib/server
